+++ src/main/java/main/dao/IDepartamentoDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import main.dto.Departamento;
 
-public interface IDepartamentoDAO extends JpaRepository<Departamento, Integer>{
+public interface IDepartamentoDAO extends JpaRepository<Departamento, Long>{
 	
 }
